@@ -1,8 +1,7 @@
-// vad.dart
+// lib/vad.dart
 
 library;
 
-export 'src/vad_handler_base.dart';
 export 'src/vad_handler.dart';
 export 'src/vad_iterator.dart';
 export 'src/vad_event.dart';
